+++ BloonsTD6 Mod Helper/ModHelper.cs
@@ -20,9 +20,9 @@ namespace BTD_Mod_Helper;
 /// </summary>
 public static class ModHelper
 {
-    internal const string WorksOnVersion = "55.0";
+    internal const string WorksOnVersion = "55.2";
     internal const string Name = "BloonsTD6 Mod Helper";
-    internal const string Version = "3.6.4";
+    internal const string Version = "3.6.5";
     internal const string RepoOwner = "gurrenm3";
     internal const string RepoName = "BTD-Mod-Helper";
     internal const string Description =